Updated App.js by Implementing Router and API Calls
Right now, I am dealing with a problem that my API Call for GetAllUsers is not working. I have done all safety checks especially I am returning a value of string for the header parameters. However, I  kept on getting the 401 Unauthorized Error. Although, If I changed the value of parameter with the exact value of the string, it worked. So I don't know how to fix this problem as of now